fix: #190 アプリを起動するとメモリリークの警告が出力される
レビュー指摘対応: winstonインスタンスの作成をstaticでWinstonLoggerImpl内に定義
