+++ src/main/inversify.logger.config.ts
@@ -1,12 +1,9 @@
 import { Container } from 'inversify';
-import winston from 'winston';
 import { TYPES } from './types';
 import { ILogger } from './services/ILogger';
 import { ILoggerFactory } from './services/ILoggerFactory';
 import { LoggerFactoryImpl } from './services/LoggerFactoryImpl';
 import { WinstonLoggerImpl } from './services/WinstonLoggerImpl';
-import { ILoggerInitializer } from './services/ILoggerInitializer';
-import { WinstonInitializerImpl } from './services/WinstonInitializerImpl';
 
 // コンテナの作成
 const container = new Container();
@@ -16,13 +13,5 @@ const container = new Container();
 // そのため、 inTransientScope で呼び出しごとにインスタンスを新しく作成することで setName が同時に呼び出されないようにする。
 container.bind<ILoggerFactory>(TYPES.LoggerFactory).to(LoggerFactoryImpl).inTransientScope();
 container.bind<ILogger>(TYPES.WinstonLogger).to(WinstonLoggerImpl).inTransientScope();
-
-// winstonのインスタンスはイベントリスナーを使用しているため、
-// 複数インスタンスを作成すると、デフォルトのリスナー上限値を超えしまうので、メモリリークの警告が出力される。
-// そのため、winstonのインスタンスを共有化し、1つのインスタンスを使いまわしてログ出力を行う。
-container
-  .bind<ILoggerInitializer<winston.Logger>>(TYPES.WinstonInitializer)
-  .to(WinstonInitializerImpl)
-  .inSingletonScope();
 
 export default container;
