--- src/main/services/ILoggerInitializer.ts
@@ -1,3 +1,0 @@
-export interface ILoggerInitializer<T> {
-  getLogger(): T;
-}
